Update Product In Bill Detail

+++ src/main/java/com/example/java4/entities/KhuyenMai.java
@@ -2,7 +2,6 @@ package com.example.java4.entities;
 
 import jakarta.persistence.*;
 
-import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.Objects;
 
@@ -13,7 +12,7 @@ public class KhuyenMai {
     private String ten;
     private Date ngayBatDau;
     private Date ngayKetThuc;
-    private BigDecimal soTienGiam;
+    private Integer soTienGiam;
     private Integer soLuong;
     private Integer trangThai;
 
@@ -65,11 +64,11 @@ public class KhuyenMai {
     }
 
     @Column(name = "SoTienGiam")
-    public BigDecimal getSoTienGiam() {
+    public Integer getSoTienGiam() {
         return soTienGiam;
     }
 
-    public void setSoTienGiam(BigDecimal soTienGiam) {
+    public void setSoTienGiam(Integer soTienGiam) {
         this.soTienGiam = soTienGiam;
     }
